Clearify error message
The previous error message was a bit misleading, as the listed files are
not the header files that are missing, but apparently the files that
were missing a header.

This patch fixes this by making the error message more clear and not
printing the debug representation of the `Vec<String>` that contains the
file pathes, but a comma-seperated list representation of these pathes.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>